Replace ws with ssl
