+++ litinit.go
@@ -2,10 +2,18 @@ package main
 
 import (
 	"bufio"
+	"crypto/rand"
+	"crypto/rsa"
+	"crypto/x509"
+	"crypto/x509/pkix"
+	"encoding/pem"
 	"io"
 	"log"
+	"math/big"
+	"net"
 	"os"
 	"path/filepath"
+	"time"
 
 	"github.com/jessevdk/go-flags"
 	"github.com/mit-dci/lit/lnutil"
@@ -29,6 +37,112 @@ func createDefaultConfigFile(destinationPath string) error {
 		return err
 	}
 	writer.Flush()
+	return nil
+}
+
+// helper function to see if a given path exists or not
+func pathExists(path string) bool {
+	_, err := os.Stat(path)
+	if err == nil {
+		return true
+	}
+	if os.IsNotExist(err) {
+		return false
+	}
+	return true
+}
+
+// GenCert() creates ad directory, and appropriate server and client keys/certs
+func GenCert() error {
+	// generate cert
+	if !pathExists("certs") {
+		log.Println("Creating certs directorty")
+		err := os.Mkdir("certs", 0775)
+		if err != nil {
+			log.Println(err)
+			return err
+		}
+	}
+	if !pathExists("certs/server.key") {
+		// if one is deleted, doesn't make sense to have the other
+		log.Println("Generating server cert")
+		err := genCertHandler("server")
+		if err != nil {
+			log.Println(err)
+			return err
+		}
+	}
+
+	if !pathExists("certs/client.key") {
+		log.Println("Generating client cert")
+		err := genCertHandler("client")
+		if err != nil {
+			log.Println(err)
+			return err
+		}
+	}
+	return nil
+}
+
+// genCertHandler creates a .pem and .key file for use by the listener and client respectively
+// Template adapted from https://gist.github.com/glennwiz/74b01bc3dc916bdd2446
+
+func genCertHandler(name string) error {
+
+	var err error
+
+	cert := x509.Certificate{
+		Subject: pkix.Name{
+			Organization: []string{"Lightning Network"},
+		},
+		NotBefore: time.Now(),
+
+		KeyUsage:              x509.KeyUsageCertSign,
+		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth},
+		BasicConstraintsValid: true,
+
+		IsCA: true,
+	}
+
+	cert.IPAddresses = append(cert.IPAddresses, net.ParseIP("127.0.0.1"))
+	cert.NotAfter = cert.NotBefore.Add(time.Duration(365) * time.Hour * 24)
+
+	priv, err := rsa.GenerateKey(rand.Reader, 2048)
+	if err != nil {
+		log.Println("Failed to generate private key:", err)
+		return err
+	}
+
+	serialNumberLimit := new(big.Int).Lsh(big.NewInt(1), 128)
+	cert.SerialNumber, err = rand.Int(rand.Reader, serialNumberLimit)
+	if err != nil {
+		log.Println("Failed to generate serial number:", err)
+		return err
+	}
+
+	derBytes, err := x509.CreateCertificate(rand.Reader, &cert, &cert, &priv.PublicKey, priv)
+	if err != nil {
+		log.Println("Failed to create certificate:", err)
+		return err
+	}
+
+	destPath := "certs/" + name + ".pem"
+	certOut, err := os.Create(destPath)
+	if err != nil {
+		log.Println("Failed to open server.pem for writing:", err)
+		return err
+	}
+	pem.Encode(certOut, &pem.Block{Type: "CERTIFICATE", Bytes: derBytes})
+	certOut.Close()
+
+	keyPath := "certs/" + name + ".key"
+	keyOut, err := os.OpenFile(keyPath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0600)
+	if err != nil {
+		log.Println("failed to open server.key for writing:", err)
+		return err
+	}
+	pem.Encode(keyOut, &pem.Block{Type: "RSA PRIVATE KEY", Bytes: x509.MarshalPKCS1PrivateKey(priv)})
+	keyOut.Close()
 	return nil
 }
 
@@ -130,6 +244,9 @@ func litSetup(conf *config) *[32]byte {
 	if err != nil {
 		log.Fatal(err)
 	}
-
+	err = GenCert()
+	if err != nil {
+		log.Fatal(err)
+	}
 	return key
 }
